+++ auto-load-next-post.php
@@ -5,7 +5,7 @@
  * Description: Increase your pageviews on your site as readers continue reading your posts scrolling down the page.
  * Author:      Sébastien Dumont
  * Author URI:  https://sebastiendumont.com
- * Version:     1.5.3
+ * Version:     1.5.5
  * Text Domain: auto-load-next-post
  * Domain Path: /languages/
  *
@@ -24,8 +24,6 @@
  * Main Auto Load Next Post Class
  *
  * The main instance of the plugin.
- *
- * @version 1.5.0
  */
 if ( ! class_exists( 'Auto_Load_Next_Post' ) ) {
 
@@ -47,7 +45,7 @@ if ( ! class_exists( 'Auto_Load_Next_Post' ) ) {
 		 * @static
 		 * @since  1.5.0
 		 */
-		public static $version = '1.5.3';
+		public static $version = '1.5.5';
 
 		/**
 		 * Integrations instance.
@@ -138,7 +136,7 @@ if ( ! class_exists( 'Auto_Load_Next_Post' ) ) {
 		 * Setup Constants
 		 *
 		 * @since   1.4.3
-		 * @version 1.5.0
+		 * @version 1.5.5
 		 * @access  private
 		 */
 		private function setup_constants() {
@@ -155,8 +153,12 @@ if ( ! class_exists( 'Auto_Load_Next_Post' ) ) {
 			$suffix       = defined( 'SCRIPT_DEBUG' ) && SCRIPT_DEBUG ? '' : '.min';
 			$debug_suffix = defined( 'ALNP_DEV_DEBUG' ) && ALNP_DEV_DEBUG ? '.dev' : '';
 
-			$this->define( 'AUTO_LOAD_NEXT_POST_SCRIPT_MODE', $suffix );
-			$this->define( 'AUTO_LOAD_NEXT_POST_DEBUG_MODE', $debug_suffix );
+			$this->define('AUTO_LOAD_NEXT_POST_SCRIPT_MODE', $suffix);
+			$this->define('AUTO_LOAD_NEXT_POST_DEBUG_MODE', $debug_suffix);
+
+			$this->define('AUTO_LOAD_NEXT_POST_STORE_URL', 'https://autoloadnextpost.com/');
+			$this->define('AUTO_LOAD_NEXT_POST_SUPPORT_URL', 'https://wordpress.org/support/plugin/auto-load-next-post');
+			$this->define('AUTO_LOAD_NEXT_POST_REVIEW_URL', 'https://wordpress.org/plugins/auto-load-next-post/#reviews');
 		} // END setup_constants()
 
 		/**
@@ -172,10 +174,6 @@ if ( ! class_exists( 'Auto_Load_Next_Post' ) ) {
 				define( $name, $value );
 			}
 		} // END define()
-
-		/*-----------------------------------------------------------------------------------*/
-		/*  Load Files                                                                       */
-		/*-----------------------------------------------------------------------------------*/
 
 		/**
 		 * Include required core files used in admin and on the frontend.
@@ -270,10 +268,6 @@ if ( ! class_exists( 'Auto_Load_Next_Post' ) ) {
 			do_action( 'auto_load_next_post_init' );
 		}
 
-		/*-----------------------------------------------------------------------------------*/
-		/*  Localization                                                                     */
-		/*-----------------------------------------------------------------------------------*/
-
 		/**
 		 * Make the plugin translation ready.
 		 *
@@ -326,10 +320,6 @@ if ( ! class_exists( 'Auto_Load_Next_Post' ) ) {
 				) );
 			} // END if is_singular() && get_post_type()
 		} // END alnp_enqueue_scripts()
-
-		/*-----------------------------------------------------------------------------------*/
-		/*  Helper Functions                                                                 */
-		/*-----------------------------------------------------------------------------------*/
 
 		/**
 		 * Checks if we are using the theme customizer.