+++ includes/admin/settings/class-alnp-settings-page.php
@@ -10,7 +10,7 @@
  * @license  GPL-2.0+
  */
 
-if ( ! defined('ABSPATH')) {
+if ( ! defined('ABSPATH') ) {
 	exit; // Exit if accessed directly.
 }